test(rust): rename rust test helper to avoid _helpers sys.modules collision
The rust test suite shipped tests/parsers/rust/_helpers.py and imported it as the
bare module `from _helpers import ...`. The swift suite already ships a
tests/parsers/swift/_helpers.py imported the same way. Under pytest's default
prepend import mode both resolve to a single sys.modules["_helpers"]; whichever
test dir is collected first wins, and the other parser's tests then run against
the WRONG helper (the other parser's build/extract), producing empty output and
~69 spurious failures (all swift stage/scanner/reachability tests). This is why
master (swift only) is green but the feature branch (swift + rust) is red.

Rename the rust helper to _rust_helpers and update the imports so the two suites
no longer share a module name. Full suite: 2526 passed / 0 failed (was 70 failed).
