+++ libs/openant-core/tests/parsers/rust/test_rust_nonnominal_impl.py
@@ -3,7 +3,7 @@ unit) must still have their methods extracted. `_bare_type_name` names only nomi
 types, so `impl Serialize for u32` was dropped entirely (every method lost)."""
 import pathlib, sys
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parent))
-from _helpers import extract  # noqa: E402
+from _rust_helpers import extract  # noqa: E402
 
 
 def test_nonnominal_self_impls_extracted(tmp_path):
